Clean up component structure, naming. Optimized structure, removed unneeded wrappers. Fixed css, styling issues. Fixed type rendering.
Added Theme support.
Added Footer
Refactored HeaderAppBar
