+++ src/App.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import App from './App';
-import AddUserBtn from "./components/header/AddUserBtn";
+import AddUserButton from "./components/layout/AddUserButton";
 
 
 // Note: running cleanup afterEach is done automatically for you in @testing-library/react@9.0.0 or higher
