+++ src/theme.js
@@ -1,0 +1,8 @@
+import {createMuiTheme} from '@material-ui/core/styles'
+
+const theme = createMuiTheme({
+    palette: {
+        primary: {500: '#467fcf'},
+    }
+})
+export default theme
